refactor(auth): migrate to better-auth library and enhance sign-out error handling
