+++ web-app/src/app/(landing)/(auth)/forgot-password/actions.ts
@@ -2,7 +2,7 @@
 
 import { z } from "zod";
 
-import { auth } from "@/lib/auth";
+import { auth } from "@/lib/better-auth/auth";
 
 const forgotPasswordSchema = z.object({
   email: z.string().email(),
