+++ web-app/src/app/api/[...all]/route.ts
@@ -1,5 +1,5 @@
 import { toNextJsHandler } from "better-auth/next-js";
 
-import { auth } from "@/lib/auth";
+import { auth } from "@/lib/better-auth/auth";
 
 export const { POST, GET } = toNextJsHandler(auth);
