+++ web-app/src/lib/better-auth/auth.client.ts
@@ -7,9 +7,6 @@ import {
 } from "better-auth/client/plugins";
 import { createAuthClient } from "better-auth/react";
 
-export type Session = typeof authClient.$Infer.Session;
-export type User = typeof authClient.$Infer.Session.user;
-
 export const authClient = createAuthClient({
   plugins: [
     organizationClient(),
